+++ crates/core/tedge_agent/src/agent.rs
@@ -1,13 +1,13 @@
 use crate::file_transfer_server::actor::FileTransferServerBuilder;
 use crate::file_transfer_server::actor::FileTransferServerConfig;
 use crate::operation_file_cache::FileCacheActorBuilder;
+use crate::operation_workflows::builder::TedgeOperationConverterBuilder;
+use crate::operation_workflows::config::OperationConfig;
 use crate::restart_manager::builder::RestartManagerBuilder;
 use crate::restart_manager::config::RestartManagerConfig;
 use crate::software_manager::builder::SoftwareManagerBuilder;
 use crate::software_manager::config::SoftwareManagerConfig;
 use crate::state_repository::state::agent_state_dir;
-use crate::tedge_operation_converter::builder::TedgeOperationConverterBuilder;
-use crate::tedge_operation_converter::config::OperationConfig;
 use crate::tedge_to_te_converter::converter::TedgetoTeConverter;
 use crate::AgentOpt;
 use crate::Capabilities;
